requirements.txt
beautifulsoup4==4.12.3
happybase==1.2.0
hdfs==2.7.3
lxml_html_clean==0.4.1
newspaper4k==0.9.2
kiwipiepy==0.19.1
pymongo==3.12.0
python-dotenv==0.21.1
requests==2.31.0
scikit-learn==1.0.2
tqdm==4.67.1
